requirements.txt
face-recognition==1.3.0
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
numpy==1.23.5
opencv-python==4.8.1.78
psycopg2-binary==2.9.10